test: extract authenticateAsTestUser() and parameterize Playwright webServer command (#466)
