+++ e2e/setup/global.setup.ts
@@ -1,8 +1,7 @@
 import "dotenv/config";
 import prisma from "@/server/prisma";
 import { FullConfig, chromium } from "@playwright/test";
-import { addMonths, getUnixTime } from "date-fns";
-import { testUser } from "../test-utils";
+import { authenticateAsTestUser, testUser } from "../test-utils";
 
 /**
  * `InvoiceVersion.invoiceId` is `onDelete: Restrict` (docs/adr/0004) — a
@@ -33,19 +32,8 @@ async function globalSetup(config: FullConfig) {
 
 	const browser = await chromium.launch();
 	const page = await browser.newPage();
-	await page.goto(baseURL!);
 
-	await page.context().addCookies([
-		{
-			name: "next-auth.session-token",
-			value: testUser.sessions.create.sessionToken,
-			domain: "localhost",
-			path: "/",
-			httpOnly: true,
-			sameSite: "Lax",
-			expires: getUnixTime(addMonths(new Date(), 1))
-		}
-	]);
+	await authenticateAsTestUser(page, baseURL!);
 
 	await page.context().storageState({ path: storageState as string });
